Updated a few spots where player_t::CurrentPlayerClass was still being saved into a BYTE instead of an integer, and send a player's class as an unsigned short in "SpawnPlayer" server commands.

+++ src/cl_main.cpp
@@ -3569,7 +3569,7 @@ void ServerCommands::SpawnPlayer::Execute()
 
 	// [BB] We may not filter coop inventory if the player changed the player class.
 	// Thus we need to keep track of the old class.
-	const BYTE oldPlayerClass = pPlayer->CurrentPlayerClass;
+	const int oldPlayerClass = pPlayer->CurrentPlayerClass;
 
 	// Set up the player class.
 	pPlayer->CurrentPlayerClass = clamp<int>( playerClass, 0, PlayerClasses.Size() -1 );
